backend/api/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0
sqlmodel>=0.0.14
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
PyJWT>=2.8,<3
email-validator>=2.2,<3
python-multipart>=0.0.9,<0.1

